factor out some types to WireNetworkInterface target
