+++ WireAuthentication/Sources/WireAuthenticationLogic/NetworkStack.swift
@@ -21,6 +21,7 @@ import WireAuthenticationAPI
 import WireFoundation
 import WireLogging
 import WireNetwork
+import WireNetworkInterface
 
 package final class NetworkStack {
 
@@ -211,7 +212,7 @@ private extension NetworkService {
             )
         )
 
-        let proxySettings: WireNetwork.ProxySettings?
+        let proxySettings: WireNetworkInterface.ProxySettings?
         if let configProxySettings = backendConfig.proxySettings {
             if configProxySettings.needsAuthentication {
                 guard let proxyCredentials else {
