+++ WireNetwork/Sources/WireNetwork/Network/URLSessionConfigurationFactory.swift
@@ -17,6 +17,7 @@
 //
 
 public import Foundation
+import WireNetworkInterface
 
 public struct URLSessionConfigurationFactory {
 
@@ -68,6 +69,35 @@ public struct URLSessionConfigurationFactory {
         }
 
         return configuration
+    }
+
+}
+
+extension ProxySettings {
+
+    func proxyDictionary() -> [AnyHashable: Any] {
+        let socksEnable = "SOCKSEnable"
+        let socksProxy = "SOCKSProxy"
+        let socksPort = "SOCKSPort"
+
+        var result: [AnyHashable: Any] = [
+            socksEnable: 1,
+            kCFProxyTypeKey: kCFProxyTypeSOCKS,
+            kCFStreamPropertySOCKSVersion: kCFStreamSocketSOCKSVersion5
+        ]
+
+        switch self {
+        case let .unauthenticated(host, port):
+            result[socksProxy] = host
+            result[socksPort] = port
+        case let .authenticated(host, port, username, password):
+            result[socksProxy] = host
+            result[socksPort] = port
+            result[kCFStreamPropertySOCKSUser] = username
+            result[kCFStreamPropertySOCKSPassword] = password
+        }
+
+        return result
     }
 
 }
