+++ WireNetwork/Sources/WireNetworkInterface/Models/ResolvedBackendMetadata.swift
