+++ WireNetwork/Tests/WireNetworkTests/Backend/PinnedKeyTests.swift
@@ -17,6 +17,7 @@
 //
 
 import XCTest
+import WireNetworkInterface
 
 @testable import WireNetwork
